requirements.txt
fastapi>=0.111
uvicorn>=0.30
python-dotenv>=1.0
python-jose[cryptography]>=3.3
pydantic>=2.5
boto3>=1.26
botocore>=1.29
mangum>=0.19
requests>=2.32
